[libgens] EEPRomI2C: Implemented the internal page cache.
The reason there's a page mask is because the I2C EEPROM chips actually
cache writes, and they don't write to EEPROM until a STOP condition
is received.

The internal cache is 256 bytes, which should be enough for all EEPROMs
up to 24C1024/24LC1025. Only the portion that is actually used by a
given EEPROM is used, though.