[gens-sdl] Initial basic SDL frontend.
Instead of trying to fix the stuttering by screwing with the Qt frontend,
let's try creating a new very basic frontend with SDL. Then, once I get
it working, we can port some of the changes over to Qt to fix that.

Currently, gens-sdl displays video as fast as possible. This will need to
be synchronized to 60 fps (or 50 fps) in order to be playable.