[gens-sdl] sdl_audio_callback(): Fix buffer underflow.
The "compensate for desynchronization" code should be commented out
because it causes audio to randomly speed up. That being said, I'm
still not sure why the audio is lagging and overflowing in some cases.
The compensation code fixes the overflow at the expense of making
the audio "jump forward" every so often.