[libgens] IO/: Splitting out all of the devices into individual classes, again.
The devices were all separate classes prior to the io-manager-rewrite
branch; however, the individual devices all had to be managed by the
frontend. This time, IoManager will still manage everything, and will
have the same public interface as before. The only difference is,
instead of having an array of IoDevice_t, it will have an array of
Device* objects.

Ports 1, 2, and EXT must be non-nullptr. For empty ports, they'll be
instances of Device.

Other ports can be nullptr if they're empty.

NOTE: Device and Device subclasses are all in the LibGens::IO namespace.
IoManager will still be in LibGens.

TODO: Update IoManager to use the new Device classes.