[libgens] VdpPalette: Added cram_addr_mask.
This will be used to mask CRAM addresses for different VDP types.
Right now, it's initialized to 0x7F (MD) and isn't changed.

CRAM 8-bit accessors: XOR with U16DATA_U8_INVERT. Not sure if
this is actually needed, but it can't hurt right now, since
we're not using 8-bit accesses.