[libgens] VdpPalette: Replace mdColorMask with m5m4bits.
M5=1,M4=0 results in only the LSBs of color components being used.
However, I suspect M5=0,M4=1 does the opposite: only the top two bits
of each color component are used, but with a 2-bit DAC instead of
simply keeping the low bit 0.

The following commits will be refactoring all of the update logic
for the various VDP modes, as well as reworking PalMode_t.