[libgens] IoColecoVision: Preliminary ColecoVision controller support.
The ColecoVision system is virtually identical to the SG-1000, except
it has a different I/O map and controller. (The SG-1000 was actually
based on the ColecoVision.)

Among other things, the controller has a keypad. The keypad is selected
by switching the states of pin 5 (+5V) and pin 8 (ground), which aren't
usually selectable by Sega hardware. Hence, an extra function, setPin58(),
was added to Device in order to control these lines. This function will
only be used when emulating the ColecoVision system.

TODO: Check pin58 on other controllers, since changing +5V/GND on most
controllers will either make it not work (SMS) or possibly damage the
controller (3BTN, 6BTN, etc.).