[libgens] Vdp: Added a special 16-color "app" mode for app-based operating systems.
Some app-based operating systems insist on using 16-color mode for everything,
so in order to fit in, Gens/GS II needs to reduce the color palette to 16 colors.

This is done by forcing a CGA palette (with inverted intensity bit, since most
MD games use lower palette indexes for primary objects) on all the active
palette lines. This breaks palette effects on most games, but since Gens/GS II
is a program, not an app, this isn't a problem.