Moved mcd_rom_db.c from LibGens to GensQt4.
The Mega CD ROM database is only used for the UI, and has no bearing
on emulation. In addition, it has hard-coded English strings, which
makes it difficult to translate.

This is an "as-is" move. The database needs to be rewritten to make
it easier to translate using Qt's translation system.