[cz80] Implemented the WZ register.
The WZ register is used for address calculations, and isn't usually
visible to the programmer. The one notable exception is the BIT n,(HL)
instruction, which leaks bits 3 and 5 of the high byte of WZ.

In order to properly emulate BIT n,(HL), the WZ register has to be
implemented. This obviously adds some overhead, but not much:

Benchmarks:
- cz80: Release build, gcc-4.9.2, i686

               |   ZEXDOC   |   ZEXALL   |
---------------+------------+------------+
 cz80, no WZ   |  73,623 ms |  72,976 ms |
 cz80, with WZ |  74,291 ms |  75,449 ms |
---------------+------------+------------+
      % change |    0.9073% |    3.3888% |

(I'm using an actual % change algorithm this time, instead of the
previous algorithms that were simply dividing one by another,
resulting in a ratio.)

The ZEXALL change is odd, since it's larger than ZEXDOC now.

WZ is now a register local variable, similar to PC, When it was
updated directly in the CPU struct, the performance was worse:

               |   ZEXDOC   |   ZEXALL   |
---------------+------------+------------+
 cz80, no WZ   |  73,623 ms |  72,976 ms |
 cz80, mem WZ  |  82,529 ms |  81,796 ms |
---------------+------------+------------+
      % change |   12.0968% |   12.0862% |

The following ZEXALL test now passes:
- bit n,<b,c,d,e,h,l,(hl),a>....  OK

These ZEXDOC tests still fail:
- cpd<r>........................  ERROR **** crc expected:a87e6cfa found:006b7379
- cpi<r>........................  ERROR **** crc expected:06deb356 found:2d8289a0

These ZEXALL tests still fail:
- cpd<r>........................  ERROR **** crc expected:134b622d found:a66e68d7
- cpi<r>........................  ERROR **** crc expected:2da42d19 found:47d7f28a