[cz80] Fix the CPIR and CPDR instructions on both ZEXDOC and ZEXALL.
CZ80 was checking for 'res == 0' in the main loop, but not when it
was determining to go to the next instruction or end execution.
This is now checked.

Note that we have to check CZ80_ZF instead of res, since res is
modified if CZ80_HF is set.

Thanks to flamewing in #SSRG for debugging this error.

All ZEXDOC and ZEXALL tests now pass!

TODO: Fix cycle counts. Several instructions have incorrect cycle timing.
 They need to be resynchronized with the official Z80 documentation.