[cz80] cz80_opxycb.inc: Inlined all XYCB opcodes.
Many of these opcodes use the exact same format but with a different
operation. To handle these, I created "UOP" macros that are specified
as a parameter for the main macro. This helps to reduce unnecessary
lines of boilerplate code.

OPXYCB_BIT_N_mXYd(): 8 parameters are needed, since jump table labels
are created using ##, and computing the labels doesn't work, since we'd
end up with completely broken labels. The switch/case version worked
fine with a single parameter and computed cases, though.