[gens-sdl] Ported the audio ringbuffer code from DGen/SDL 1.33.
This still needs some cleanup, e.g. audio frequency and channels
are currently hard-coded, but it appears to work a lot better than
the code I implemented before.

Note that this isn't a true ringbuffer. It works similarly to
the old SDL shifting code, but this version actually seems to
work well.

Sidenote: DGen/SDL is GPL-licensed, so we can do this.
I'll probably send my CZ80 changes back to DGen later.