[gens-sdl] VBackend: New VBackend base class.
This class only has abstract functions and basic management for the internal
framebuffer. I was going to make set_video_source a base function instead
of abstract, but the SDL version needs to do some stuff to use the MdFb as
an SDL surface, so it's kept as abstract for now.