[gens-sdl] SdlGLBackend: Initial OpenGL backend.
Based on gens-qt4's GLBackend.

VBackend::update(): Added a parameter to indicate if MdFb is dirty.
This is only used by SdlGLBackend; SdlSWBackend always has to blit
the texture, since it's copying directly to the screen buffer instead
of to a texture object.

SdlHandler: Create an SdlGLBackend instead of SdlSWBackend.

TODO:
- Split the GL code into GLBackend, and make SdlGLBackend a simple
  SDL wrapper around it.
- Port more functionality:
  - Properties (aspect ratio constraint, stretch mode)
  - Onscreen display
  - Preview image
- Enable window resizing.
- Add a command line parameter to select between SdlSWBackend and
  SdlGLBackend.