[gens-sdl] VBackend: Added a dirty flag.
This flag is set when a property is changed, and is cleared in
derived classes' update() functions.

SdlHandler: New function update_video_paused() that only updates
video if VBackend is dirty.

gens-sdl.cpp: If emulation is paused, call update_video_paused()
to ensure that VBackend property changes take effect immediately.

Stretch mode changes now take effect immediately when emulation is
paused, instead of waiting for emulation to be unpaused.