[libgens] Initial Pico page register and controller functionality.
Buttons are read from VIRTPORT_1 and are remapped such that
button 4 == the "red" button, and button 5 == the pen button.

Page register is currently mapped to PgUp and PgDn in gens-sdl.
Note that some games are "missing" pages. For example,
"Huckle & Lowly's Busiest Day Ever" has pages 1-4, but the paint
section at the back of the book is actually page 6. Switching to
page 5 results in the currrent page being reloaded.

TODO:
- When actual page image support is added, skip over pages that
  aren't present in the page image archive.
- Check how many sensors are present on the Pico. I'm thinking 7,
  but will allow 8 for now.
- Add page register controls to KeyManager instead of hard-coding
  PgUp/PgDn in gens-sdl.