[gens-sdl] Moved the Pico page control buttons to the controller.
A = previous page
C = next page

TODO: Create an IoPico controller instead of overloading Io3BTN
in order to make the page register stuff less of a hack, since
pressing A or C requires the page register to be adjusted, and
it should only be adjusted one time per press.

[libgens] IoManager: Don't adjust the bits. This makes the Start
button equivalent to the pen button, and B is still the Red button.