[libgens] Preliminary IoPico controller class.
It's basically a clone of Io2BTN right now, but it will support
page control and pen position as well.

There's also some workarounds to disable the tristate controls
inherent to IoDevice. Pico controls are directly connected to
input-only lines, so there's no tristate.

Also, the pen button is mapped to bit 7, which is normally a latched
register. Work around this by copying deviceData to mdData.

TODO: Don't allow use of IoPico if we're not emulating Pico.