[libgens] IoPico: Added page control register functions.
If Page Up or Page Down is pressed, the page number will be changed.
TODO: Show an OSD message.

IO::Device: Added a buttons_prev field. This field contains the
buttons value prior to update(). Used by IoPico to determine if a
Page button was just pressed now or if the user is holding it down.

TODO:
- Add an "unusable page" function.
- Switch to IoPico.