[libgens] IoManager: Moved page control functionality to IoPico.
picoNextPage() and picoPrevPage() were removed.

picoCurPage() and picoGetPageRegister() now check if the controller
in VIRTPORT_1 is IoPico; if it is, it calls the appropriate functions
for that device.

IoPico:
- Fixed incorrect parentheses for PgDn/PgUp checking.
- Marked some virtual functions as final.

[gens-sdl] Created a new keyMap for Pico controllers.

This maps PgDn and PgUp to the PgDn and PgUp keys. Also, it uses
IOT_PICO for the Pico controller instead of IOT_6BTN.

Removed the page control functions from gens-sdl.cpp, since it's
now implemented in IoPico.