[libgens] Rom: Added support for SMD-format ROMs.
SMD "split" format is partially supported in that the split ROM
will be loaded, but additional segments will not be loaded.
I don't know if/when I'll implement proper split ROM loading.

Reworked header parsing such that DetectSystem() always uses a
deinterleaved copy of the header. This reduces the amount of
detection code, since the SMD/MGD-specific code can be removed.

TODO:
- MGD support. This requires reading the second half of the ROM,
  which is not currently doable due to the design of Decompressor.
- Add a function to determine the overhead required by the read buffer
  instead of hard-coding 512 bytes in RomCartridgeMD.
- Eliminate the overhead by adding a 'seek' parameter in Decompressor.