[libgens] Timing_win32.cpp: Fixed a rounding error that caused screwy framerates.
On Wine, the performance frequency is emulated as 10,000,000 Hz.

On my Windows VM and a Windows 7 test PC, this wasn't the case:
- Windows XP VM in VirtualBox 5.0: 3,579,545 Hz -> "50 fps" (with fast music)
- Windows 7 PC: 3,247,099 - 3,247,109 Hz -> "56 fps"

This was caused by rounding errors, since we were dividing the frequency
by 1,000,000 first. To fix this, multiply the counter by 1,000,000,
then divide by the frequency.

FIXME: Prevent overflow, which will happen after slightly more than 7 months.