[gens-qt4] GensWindow_menu_init.cpp: New menu initialization code.
initMenuBar(): This function is now used to initialize some stuff
that can't be set in Qt Designer.

updateMenuBarVisibility(): This function has the menu bar visibility
code that used to be in initMenuBar(). Note that the menu bar cannot
currently be re-enabled once it's hidden, since neither the hotkey
nor the right-click menu are working.