[gens-qt4] GensWindow: Added slots and shortcut keys for "Stretch Mode" and "Region".
The shortcut key is the only way to activate these, since clicking the
menu item results in the submenu opening, and doesn't trigger anything.

Also, the QAction containing the menu has to be connected, not the
menu itself. The QMenu's triggered() action occurs if any submenu
item is triggered.

NOTE: QMenu::triggered() could be used instead of a QSignalMapper,
but that requires checking the QAction parameter, so there's really
no benefit in doing that.