[gens-qt4] GensWindow: Reimplemented popup menu support.
When the menu bar is hidden, the popup menu is activated on right-click.
The popup menu contains the menus created by Qt Designer, so they're
always in sync with the regular menu bar.

TODO: If emulating the Sega Mega Mouse, add a hotkey to capture
the mouse to allow use of right-click.
the commit message for your changes. Lines starting