[libzomg] img_data.h: New fields for pHYs (aspect ratio).
This will be used to save the correct aspect ratio to the PNG image.

I don't think any image viewers actually use it, but we might as well
save the value since the chunk is there.

PngWriter, PngReader: Updated to save/load the pHYs chunk.

PngWriter: If phys_x == 0 || phys_y == 0, pHYs is not saved.

PngReader: If pHYs doesn't exist, phys_x and phys_y are set to 0.
Also, the unit type is ignored.