[libgens] Util/Screenshot.cpp: New static class.
This class takes a screenshot given an MdFb and Rom, plus a filename.
All of the metadata is collected here.

Based on the code from gens-sdl, but with a few improvements.
Notably, the system ID is determined by checking the Rom object's
sysId property.

[gens-sdl] Config::doScreenShot(): Use LibGens::Screenshot.

TODO: Use LibGens::Screenshot in the other four places
that take screenshots.