[gens-qt4] GensWindow_menu_sync.cpp: Initial menu sync code.
Currently, synchronization is implemented for stateChange() and
for four properties that have submenus with multiple options.

Changed the toggled() slots to triggered(). toggled() is emitted
if the 'checked' property is changed either by the user or by
the program, whereas triggered() is only emitted if the action
is selected by the user. This makes it unnecessary to "lock" the
QActions when synchronizing states.

TODO: Implement syncAll(), syncRecent(), and syncShowMenuBar().