[gens-qt4] RecentRomsMenu: Removed d->setRecentRoms().
Moved the code to the main setRecentRoms() function.

There's no point in having the code be in a private class function
while the public class function is simply a wrapper if the function
is specifically for external users and isn't used by the class itself.

Fixed the slot names. Not only were they missing "_slot", the
recentRomsUpdated slots had a lowercase 'u'. This wasn't noticed
before because this function isn't actually used anywhere right now.
(If I upgraded to Qt 5 and used the new PointerToMemberFunction syntax,
it would have been caught at compile time...)