[gens-qt4] GensMenuShortcuts: Initial menu shortcuts handler.
Action names are mapped to the names used in the settings file.

This works for all of the normal menu classes.
TODO: QMenu for non-menu items.

Note that shortcuts in the UI file override these shortcuts
when retranslateUi() is called. To work around that, we call
updateActions() after retranslateUI(). (TODO: Remove the
shortcuts in the UI file?)

TODO: Split into two classes:
- Config class for the actual configuration (one per process).
- Actions class for the UI (one per window).