[libzomg] Zomg.hpp: Marked all virtual functions as final.
I'm pretty sure I won't be implementing any subclasses of Zomg.

If I were to implement a Genecyst/Gens savestate loader, it'd be a
direct subclass of ZomgBase, since the file is basically a blob
and has no structure within the file itself, whereas ZOMG is a
Zip file and has a central directory.