[libW32U] Moved gens-sdl's do_wmain() into libW32U.
The function is now W32U_GetArgvU(), and it's located in
libW32U/W32U_argv.h.

NOTE: W32U_argv.h is not included by W32U_mini.h, since it should
only be used by main().

NOTE 2: MSVC 2010 only supports C89, so in order to get it to compile
with MSVC 2010, I had to rearrange the code such that all of the
variables are declared at the top of each block.