[gens-sdl] OsdGL: Added a message fade-out effect.
When the message timer expires, it will fade out for 250ms.

This is implemented by not removing the message until its endTime
plus the fadeOutTime is passed; however, once endTime passes, the
alpha value of the message is decreased from 1.0f to 0.0f linearly
over the fade-out period.

Changed the glColor4f() parameters for resetting the GL state from
1.0 to 1.0f. No actual differences, but since this function takes
floats, not doubles, we might as well give it floats.