[gens-sdl] Don't reupload the texture if the window is exposed while emulation is paused.
Minor optimization that should help to reduce CPU usage a bit
if the window is exposed while paused. Note that this only happens
if a compositing window manager is not in use.

SdlHandler::update_video_paused(): New parameter 'force'.
If true, the VBackend is updated with fb_dirty = false, which redraws
everything to the screen but without reuploading the texture.

gens-sdl.cpp::run(): Call sdlHandler->update_video_paused() with
the 'exposed' flag as the 'force' parameter.