[gens-sdl] OsdGL: Added a display offset property to adjust for aspect ratios.
This is needed to ensure that the OSD is always drawn within
the framebuffer area.

OsdGL::draw(): Adjust the projection for the display offsets.

GLBackendPrivate::recalcAspectRatio(): Set the OSD's display offsets.

TODO: If aspect ratio constraints are disabled, maybe the visible
area should be center-aligned instead of stretching it, or the
visible area should be enlarged.