[libzomg] Metadata_win32.cpp: New function getOSVersion_8().
This function reads version information from the registry instead of
using GetVersionEx(). This is needed on Windows 8.1 and later because
they always claim to be Windows 8 unless the program has a special
manifest, and since we're not targetting Windows 8+, we won't be
including that manifest.

This method seems to work for Windows XP and later, though I'm only going
to use it if the OS version is Windows 8 or later. Older versions will
use the GetVersionEx() method.