[libW32U] Added an option to disable support for ANSI Windows.
MSVC 2008 and later don't support targetting versions of Windows earlier
than NT 5.0 (Windows 2000). Therefore, if we're using MSVC 2008 or later,
disable ANSI support unconditionally, since it's not needed.

For older versions of MSVC (which probably won't work anyway due to
missing stdint.h) and for MinGW-w64, ANSI Windows support is determined
using the CMake option -DENABLE_ANSI_WINDOWS. If enabled, the minimum
Windows version is reduced to 4.00. If disabled, the minimum Windows
version is kept at 5.00. This includes both the subsystem and the
_WIN32_WINNT macro.

NOTE: If compiling for 64-bit Windows (which I'm not supporting, but may
be possible), the minimum subsystem version is always 5.02, which
indicates Windows Server 2003 and Windows XP x64 Edition. There was
never any release of 64-bit ANSI Windows, so 64-bit builds always
have ANSI support disabled.