[libW32U] W32U_shlobj.c: New file for shlobj.h functions.
The first function is SHGetSpecialFolderPathU(). Note that the underlying
*W and *A functions take a string buffer that must be MAX_PATH length,
so lpszPath should be at least MAX_PATH*3 to prevent issues caused by
UTF-16 to UTF-8 expansion.

To reduce issues, I added a cbPath parameter that specifies the size of
lpszPath. If cbPath isn't big enough, the function will return an error.

NOTE: We're using SHGetSpecialFolderPath() instead of SHGetFolderPath()
now due to some weirdness with how this function is implemented. On
Windows 2000 and later, this function is in shell32.dll. On Win9x,
this function is in shfolder.dll, which isn't part of the OS.
SHGetSpecialFolderPath() is available on Win95+IE4 and Win98
as part of shell32.

Since we don't need any of the extra functionality of SHGetFolderPath(),
we can switch to SHGetSpecialFolderPath() for Win9x compatibility without
having to include shfolder.dll. (Windows 2000 and later includes an
shfolder.dll shim that redirects to shell32.dll, so that approach could
have worked, but I chose not to use it.)

[gens-sdl] Config.cpp::getConfigDir(): Use SHGetSpecialFolderPathU().