[libcompat] reentrant.h: Added a placeholder for HAVE_GETPWUID_R.
Implementing a wrapper for getpwuid_r() requires copying all of the
fields to a temporary buffer. This is annoying to do, and since most
modern Unix and Linux systems have getpwuid_r(), I don't think I need
to implement it right now.

[gens-sdl] Config.cpp::getConfigDir(): Removed the getpwuid() path.
Always use getpwuid_r(), since we're assuming the wrapper will be there
if the system doesn't have it.