Added CMake macros for Large File Support.
On MSVC, we're using _fseeki64() and _ftelli64(). fseeko() and ftello()
aren't implemented and likely never will be implemented.

On other systems (including Linux and MinGW-w64), we check if fseeko()
and ftello() are 64-bit. If they are, then we're done. If not, we add
LFS macros and then check again. If that works, we append the LFS macros
to CFLAGS/CXXFLAGS.

TODO:
- Use _fseeki64() and _ftelli64() on MinGW-w64 to avoid the use of
  wrapper functions?
- Find any remaining users of fseek() and ftell() and switch them over
  to fseeko() and ftello() (or the equivalent MSVC version).