Split MSVC-specific C99 compatibility macros from c++11-compat.h into msvc-c99-compat.h.
These macros are for functions that exist in MSVCRT, but have a different
name compared to POSIX or C99, e.g. _snprintf() instead of snprintf().

Moved strtoll() and strtoull() from W32U_libc.h to msvc-c99-compat.h.
W32U is for functions that need character set conversion, not for
functions that simply have a different name.