[gens-sdl] VBackend: Implemented the paused effect.
VBackend::setPausedEffect() immediately enables the effect.
VBackend does not maintain the emulator's 'paused' state, so the
caller should make sure it only sets it if both the emulator is
paused and the user enabled the paused effect.

VBackend now has a "force FB dirty" flag that forces a reupload
of the framebuffer. This might not be necessary if shaders are
in use for e.g. Fast Blur.

VBackend also has an internal framebuffer that's used to apply
effects. This internal framebuffer is only allocated if it's
needed for an effect.

gens-sdl.cpp: Set the paused effect when emulation is paused.