[libgens] FastBlur.cpp: Fix pixel count and starting offset.
MdFb::fb16() and MdFb::fb32() return pointers to the first visible pixel
in the framebuffer. Hence, we don't need to adjust the mdScreen pointer
to compensate for the 8px left border. The adjustment caused the first
8 pixels of the framebuffer to not get blurred. (This probably isn't too
noticeable, but it's still an error.)

Removed the hard-coded pixel count and changed it to a parameter.

Added asserts to the MMX code to ensure that pxCount is a multiple of
4 (15/16-bit) or 2 (32-bit).