[libgens] cpuflags.c: Added MSVC support.
Inline assembler is used for some i386-only stuff.

The MSVC __cpuid() intrinsic is used for actually executing the 'cpuid'
instruction. As such, our cpuid macro is now CPUID(). Note that the
intrinsic uses an array as a return value instead of individual registers.
MSVC optimizes it away in the release build, so that's not a problem.