[cmake] platform/win32-gcc.cmake: Generate and keep the .reloc section.
This is required for proper ASLR support on Windows Vista and later.

Note that there's a bug in GNU `ld` (2.25.1 and earlier; possibly
later versions, assuming they don't bother to fix it) where enabling
--pic-executable causes it to forget the entry point. We have to
manually specify _mainCRTStartup (or mainCRTStartup on 64-bit)
in order to get the executable to work.

Based on a libav patch:
https://lists.libav.org/pipermail/libav-devel/2014-October/063871.html