[libgens/tests] AudioWriteTest: Added support for testing flag variants.
This allows us to test no-flag, MMX, and SSE2 in the same executable.
It works by changing the CPU_Flags global variable, which is used by
SoundMgr to determine which code to run.

If a CPU doesn't support MMX or SSE2, it will fail the test with an
error message indicating it's not supported.