[gens-qt4] GensPortAudio: Use SoundMgr's write functions.
Also a bunch of various cleanups.

TODO: GensPortAudio uses a (poorly-designed) ring buffer.
We have to use a bounce buffer in order to use SoundMgr's write
functions, since SoundMgr expects 16-byte alignment for SSE2.

This class will be thrown out once we port the new SDL audio code
from gens-sdl to gens-qt4 later, so it's not a big deal.