CheckPopt.cmake: Check for libpopt.
The system popt will be used if available. If it isn't, the internal
copy of popt will be compiled and used instead.

Note that unlike zlib, minizip, libpng, and GLEW, we're going to check
for the system popt on Windows. I should probably use the system
versions for the other libraries on Windows, but maybe later.