[gens-sdl] OsdFont.h: New OsdFont struct.
Reworked the C64 and VGA fonts to use the OsdFont struct, which contains
the character data, width, and height.

OsdGL: Use the width and height from the specified OsdFont. This allows
us to switch between the VGA and C64 fonts by simply changing the struct
pointer. Note that it's currently hard-coded, but I could possibly add
a command line option for gens-sdl later.