[libcompat] byteswap.h: Change array byteswap arguments to uint16_t* and uint32_t*.
We're usually byteswapping entire memory blocks, so they're pretty much
guaranteed to be 16-bit or 32-bit aligned. Let's enforce it so we can
use optimized code in the generic version.

byteswap.c: Use __swab16() and __swab32(). (copied from byteswap_x86.c.)

Added assertions to ensure the pointers are actually aligned.
These assertions will be compiled out in the release build. On x86/amd64,
this won't be a problem. On PowerPC and ARM, they'll crash if the blocks
end up not being aligned.

ByteswapTest, ByteswapTest_benchmark: Cast the array to uint16_t*
or uint32_t* where applicable.