[libcompat] cpuflags.c: Added CPU vendor ID and full name.
These fields are publicly retrievable with two functions:
- LibCompat_GetCPUVendorID()
- LibCompat_GetCPUFullName()

If the fields are empty, the functions will return NULL.

LibCompat_GetCPUFlags() now saves the vendor ID and retrieves
the brand string if it's available.

TODO: Eliminate extra spaces in the brand string. On my ThinkPad T60p,
the brand string shows up as:
- "Intel(R) Core(TM)2 CPU         T7200  @ 2.00GHz"

I suspect the spaces are there to make it easier to parse the
specific model and frequency, but it gets annoying when using
the brand string as metadata.