[cmake] devkitPPC.RVL.cmake: More fixes.
Adding libraries to LDFLAGS worked for libogc, but fails miserably
for other libraries due to -Wl,--as-needed. If a library's symbols
aren't immediately needed, the library is removed, which can result
in undefined reference errors.

Fixed the libogc check. Check for libogc.a, not libogc's directory.
(libogc's directory is checked separately.)

Add the libraries, in reverse order, to CMAKE_C_STANDARD_LIBRARIES
and CMAKE_CXX_STANDARD_LIBRARIES. This variable adds the libraries
*after* the object files, which fixes the problem.

Added the BTE and WIIUSE library. I'm going to add a prompt to
gtest_main.cpp for the user to press the HOME button to exit.
This will be useful for libzomg's PrintMetadata test, since the
test would otherwise exit immediately after it displays the results.