[libzomg] Metadata_ogc.cpp: New Metadata implementation for libogc.
On Wii and vWii (Wii U), it gets the Wii System Menu version
and IOS version, and uses that as OS. On GameCube, it just
prints "GCN".

CPU is currently hard-coded based on target. On HW_RVL, it prints
"Broadway" or "Espresso", depending on if it's running on a real Wii
or on a Wii U in Wii Mode (vWii).

TODO:
- Add the libogc version number.
- Can the CPU frequency be obtained programmatically instead of
  hard-coding it?