[cmake] SplitDebugInformation.cmake: Use CMAKE_OBJCOPY and CMAKE_STRIP.
FIND_PROGRAM(OBJCOPY) finds the system objcopy, which fails when
cross-compiling for e.g. PowerPC. CMAKE_OBJCOPY is set to the
correct cross-compile version, so use that.

Use 'strip' instead of 'objcopy --strip-debug'. objcopy doesn't
remove the .symtab or .strtab sections, which can take up a good
amount of space. They're also copied to the .debug file, so there's
no reason to keep them in the main executable.