[libgens/tests] PausedEffectTest: Split image loading into a separate function.
We're going to use parameterized tests for CPU flags, not color depth,
so in order to selectively load the 15/16/32 images, we have to use
a separate init() function.

do32bit_1FB(), do32bit_2FB(): Call init() using ASSERT_NO_FATAL_FAILURE().
That way, if an assertion occurs in init(), it will propagate, and the
unit test will stop running.