[gens-sdl] Split the emulation loop and "Crazy" Effect loop into separate files.
All of the functions are currently in the GensSdl namespace without
any enclosing classes.

EmuLoop.cpp has its own processSdlEvent_emuLoop() function to handle keys
that aren't needed for the "Crazy" Effect, e.g. savestates. If an event
isn't handled by EmuLoop, gens-sdl.cpp's processSdlEvent_common() will
try processing it. If that doesn't handle it, the key will be sent
to the KeyManager.

TODO: Convert EmuLoop and CrazyEffectLoop to classes with a common
base class.