[gens-sdl] Preliminary absolute mouse position handling.
doAbsMouseMove() emulates a 1280x240 tablet. This allows for both
320x240 (Sega Pico) and 256x192 (SMS graphics tablets) to be handled
with a single division. (/4 for Pico, /5 for SMS.)

VBackend: New function translateAbsMouseCoords(). This is located
in VBackend because it has access to the SDL window. Also, when
using SdlSWBackend, SDL automatically scales the mouse coordinates,
so we simply have to blank it out if it's out of range. This is in
contrast to SdlGLBackend, where we have to do scaling manually.

TODO: Implement SdlSWBackend::translateAbsMouseCoords().