[libgens] IoManager: Preliminary Pico tablet support.
This commit adds absolute coordinates to the Device class, plus functions
to update the coordinates. The coordinates must be scaled to 1280x240.

1280 was chosen as the horizontal width in order to support 320px devices
as well as 256px devices. There were a few graphics tablets for the
Sega Master System that had a 256px width, so this will help to support
those devices later.

TODO: Get the mouse input from the UI and forward it to libgenskeys,
which will need to be updated to support absolute coordinates in
addition to buttons.