[libgens] Make sure all files that use GENS_ENABLE_EMULATION #include <config.libgens.h>.
Vectorman 2 crashes on the first level. I bisected this to commit
1a8b524d901c9934d2fa572cc29fc93adc2a112d:

Initial support for no-emulation builds on non-x86_32 platforms.

Adding the #includes didn't seem to fix it, so there's some other issue
with the changes that causes it to crash.

M68K.hpp, Z80.hpp: Move the inline functions out of the class definition.