[extlib] CMakeLists.txt: Check CMAKE_CL_64 for MSVC builds.
On one of my test systems, building using MSVC 2010's vcvars32.bat
for 32-bit resulted in CMAKE_SIZEOF_VOID_P == 8, since the system
was 64-bit. This may be caused by a CMake upgrade; previously,
this system had cmake-2.8.4, but now it has cmake-3.4.1.

CMAKE_CL_64 == 0 on this system, so use that for MSVC builds.