[libgens] Rom: Use Archive instead of Decompressor.
TODO: Port DcMemFake to the Archive subsystem. For now, it's disabled.
Note that DcMemFake is only used by test suites.

Note that only uncompressed ROMs will work after this change. Also, if
you load a ROM image that's an odd number of bytes, an assertion will
occur due to the byteswapping code.