[libgens] RomCartridgeMD, TmssReg: Prevent an assertion with odd-sized ROMs.
If the ROM has an odd size, add 1 to the size before byteswapping.
The final byte will end up being byteswapped with 0 (RomCartridgeMD)
or 0xFF (TmssReg).

TODO: Use 0 or 0xFF consistently.

Also, if the ROM buffer is larger than the actual size of the ROM, clear
the rest of the buffer after loading. Standard malloc() (and its aligned
variants) isn't guaranteed to clear the memory after allocating, though
it sure seems like it's cleared every time I've tried it. (Could be a
zero-page optimization on my system.)