[libgens] Zip: Ported DcZip to the Archive subsystem.
Other than changes to get it working with the Archive class,
this is basically the same as the original.

One notable change is that getFileInfo() now uses a 4096-byte
buffer for the ROM filename instead of a 256-byte buffer.
This should help with ROM files that have very long names
and/or lots of Unicode characters.