[libcompat] Moved iowin32u.c from MiniZip to libcompat/W32U/.
This file is not part of MiniZip; it's a custom I/O handler
I wrote that converts UTF-8 filenames to ANSI or UTF-16.

TODO: Use W32U functions for parts of iowin32u.

Updated everything that used iowin32u to reflect this change.

Note that minizip_iowin32u.c is compiled into a separate library,
libcompat_W32U_minizip.a. This is to prevent cross-contamination
with libcompat, since we don't want to link in MiniZip if it isn't
needed by anything.